Better structure storage factory module
